verifier bdd

+++ api/controller/challengeController.php
@@ -3,21 +3,25 @@
 require_once("model/challenge.php");
 
 class challengeController {
-	
-	$user=new User();
+	// $user=new User();
 
 	// Informations/affichage d'un challenge
 	
     public static function infos($id){
             
 		$challenge = Challenge::challenge_exists($id);
+		if (!$challenge) {
+		
+			echo(json_encode(["code" => 0,"message" => "Error : challenge does not exist"]));
+			exit();
+		}
         echo(json_encode($challenge->toArray()));
     }
 	
 	// Ajouter un challenge à la BDD
 	public static function add_challenge() {
 		
-		global $user;
+		//global $user;
 		// if ($user->is_connected(MODERATEUR)) {
 			$title=(!empty($_POST['title']))? $_POST['title']:"";
 			$desc=(!empty($_FILES['desc']))? $_FILES['desc']:"";
@@ -53,7 +57,8 @@ class challengeController {
 			$date_start=(!empty($_POST['date_start']))? $_POST['date_start']:"";
 			$date_stop=(!empty($_POST['date_stop']))? $_POST['date_stop']:"";
 			global $db;
-			$query=$db->prepare('UPDATE thechallenger.challenge SET title=:title,desc=:desc,datestart=:date_start,datestop=:date_stop WHERE id=:idchallenge');
+			$query=$db->prepare('UPDATE thechallenger.challenge SET title=:title,description=:desc,datestart=:date_start,datestop=:date_stop WHERE id=:idchallenge');
+			$query->bindParam(':idchallenge',$id,PDO::PARAM_INT);
 			$query->bindParam(':title',$title,PDO::PARAM_STR);
 			$query->bindParam(':desc',$desc,PDO::PARAM_STR);
 			$query->bindParam(':date_start',$date,PDO::PARAM_STR);
@@ -73,11 +78,16 @@ class challengeController {
 			echo(json_encode(["code" => 0,"message" => "Error : challenge does not exist"]));
 			exit();
 		}
-		$query=$db->prepare('DELETE FROM thechallenger.challenge WHERE id=:idchallenge');
-        $query->bindParam(':idchallenge',$idchallenge,PDO::PARAM_INT);
-        $query->execute();
-        $query->CloseCursor();
-		echo(json_encode(["code" => 1,"message" => "Success : challenge deleted"]));
+		global $user;
+		if ($user->is_connected(MODERATEUR)) {
+		
+		global $db;
+			$query=$db->prepare('DELETE FROM thechallenger.challenge WHERE id=:idchallenge');
+			$query->bindParam(':idchallenge',$idchallenge,PDO::PARAM_INT);
+			$query->execute();
+			$query->CloseCursor();
+			echo(json_encode(["code" => 1,"message" => "Success : challenge deleted"]));
+		}
 	}
 
 	public static function add_post() {
@@ -86,6 +96,7 @@ class challengeController {
 	}
 
 	// Envoyer tous les posts d'un challenge
+
 	public static function show_posts($idchallenge) {
 		
 		$challenge = Challenge::challenge_exists($idchallenge);
